Very basic, but working initial version
Add a new module "partitionq" based on the "partition" module of
upstream calamares source code. The upstream module cannot be used, as
it's not QML based. And because calamares' cryptsetup code uses a
keyfile and relies on the initramfs being encrypted in grub2 (which is
not the case with postmarketOS).

With this initial version, it is not possible to select the device or
partition that is the installation target. Instead, pmbootstrap will
reserve an empty partition right before the installer partition, so we
can install there and let the initramfs delete the install partition at
first boot into the installed system and re-claim the installer
partition's space.

I suggest we attempt to make this "partitionq" compatible with the
features of "partition" over time (but with a nice mobile-friendly
interface) and then upstream it into calamares.

The branding is the default branding, to be adjusted further.