welcomeq-pmos: new postmarketOS welcome screen
Add a new module "welcomeq-pmos", which displays the postmarketOS logo
and some information about what is going to be installated (arch,
device, UI, version). That information is read from a config file, which
ondev-prepare shall write in the future.

The artwork was created by Martijn Braam, and was taken from:
https://gitlab.com/postmarketOS/artwork/-/blob/master/tshirt/render-transparent.png

Related: #11