New module usersq / major refactoring
Add a new usersq module, which asks for the user's password. Since Phosh
and Plasma Mobile require a numeric password for their lockscreen
implementation, the installer asks for a numeric password. Ask the user
if they want to add an additional user (with a strong password) for SSH
access (and advise to replace password auth with key authentication once
the user's SSH key is installed). Provide the username / password dialog
to add the user.

Update the partitionq module, to make full disk encryption optional, and
to add a final confirmation screen before the installation begins.

Add input validation to all inputs (e.g. verifying that only characters
are used for the FDE password, that can be typed in with osk-sdl, make
sure that the username is valid and not reserved).

Add a navigation bar to all modules, at the top, with a back button and
title (similar to the UX of Phosh). Going back to previous screens and
modules is important, if the user changes their mind at any point in the
installation process before they hit the final confirmation button.

Refactor the code to use Calamares' jobs API instead of just directly
starting processes from the partitionq module. Usersq also has its own
job now. Besides more elegant code, this makes the UI not hang at the
end of partitionq, the commands are properly executed in the background
now while the progressbar of the installation is displayed.

Replace the default branding with the same postmarketOS 3D logo from
welcomeq. (Maybe we can add something more sophisticated in the future.)

Update ondev-prepare to properly generate the welceomeq config from
pmbootstrap data and /etc/deviceinfo.

This has become one mega-commit, as a lot of code had to be refactored.
Since postmarketos-ondev is still early in development, this should not
be a big problem - but sorry nevertheless, if somebody is trying to read
the diff.

Closes: #3, #5, #7, #11