userq: add initial version of user configuration
Create a new module, which provides the following screens:
* username
* lockscreen pin
* ssh server: enable or disable?
* ssh credentials

Inputs and basic navigation between screens is working. Most input is
properly validated, but see remaining TODOs/FIXMEs. The entered values
are not used yet, this will come in a follow-up patch. Config.{cpp,h} are
copied from welcomeq-pmos and need to be adjusted in a future patch as
well.

Related: #3