partitionq: use cryptsetup cipher from pmbootstrap
The default cryptsetup cipher in pmbootstrap is being changed to
aes-xts-plain64, and can be overridden with "pmbootstrap --cipher".
pmbootstrap passes this as ONDEV_CIPHER to ondev-prepare.sh, so pass it
down to the actual cryptsetup luksFormat command. This can be made more
elegant, once partitionq is actually using a config file.