Workaround for virtual keyboard bug: merge modules
The QT virtualkeyboard would only work reliably if the modules happened
to be loaded in a certain order. Otherwise, it would show the wrong
keyboard (i.e. qwerty keyboard instead of numpad for pin prompt) and
pressing the keys would not work. It would write "input method not set"
to the terminal. This is QTBUG-80281 in qtvirtualkeyboard.

Today I've spend multiple hours trying to find a reasonable solution for
this (and not for the first time, also Bhushan looked into it quite a
bit). The QML files are loaded in parallel, so I had patched calamares to
load them with "PreferSyncronous". But then the virtualkeyboard would
_always_ fail - even if just one QML file is enabled.

The only way I could make the virtual keyboard work reliably, is using
the parallel QML loading (as in unmodified Calamares) and enabling just
one module. We need this installer to be working properly in the
upcoming factory image, so my workaround for now is to put everything
into one module.

I've verified that the bug is gone with this patch, once in QEMU, and
10x (!) on the pinephone.

Related: #18