Support partitions in supplied rootfs image
Make it possible to install the image, if it contains a partition header
and a boot and root partition. This is the format, that most existing OS
images have, so let's support that directly. After finished
installation, we'll need to dd the boot partition over the installer OS'
boot partition (future patch).