config: when done: pmOS_inst_boot -> pmOS_boot
Support another edge case that came apparent when testing with QEMU: if
a previous installation was done to the eMMC, and then the SD card is
plugged with the installer OS, it needs to be able to find the right
boot partition (the one from SD, not from eMMC). Therefore the one in
the installerOS is now named pmOS_inst_boot and preferred by the
initramfs, and after install it is pmOS_boot again.

This patch requires the related pmbootstrap change. I'll make sure that
we release both at the same time, see the roll out plan in issue 22.